requirements.txt
astropy
astropy_healpix
numpy
Comet
pytest
git+https://github.com/ste616/atca-rapid-response-api.git#subdirectory=python
ligo.skymap
mhealpy
pre-commit